backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.11",
    "django-admin>=2.0.2",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
    "gunicorn>=23.0.0",
    "whitenoise>=6.9.0",
    "dj-database-url>=2.3.0",
]
